chore: rework oauth2 security module
