+++ src/main/java/tech/jhipster/lite/generator/server/springboot/mvc/security/oauth2/account/domain/OAuth2AccountModuleFactory.java
@@ -1,4 +1,4 @@
-package tech.jhipster.lite.generator.server.springboot.mvc.security.oauth2.domain;
+package tech.jhipster.lite.generator.server.springboot.mvc.security.oauth2.account.domain;
 
 import static tech.jhipster.lite.module.domain.JHipsterModule.*;
