+++ src/main/java/tech/jhipster/lite/generator/server/springboot/mvc/security/oauth2/core/infrastructure/primary/OAuth2ModuleConfiguration.java
@@ -1,8 +1,8 @@
-package tech.jhipster.lite.generator.server.springboot.mvc.security.oauth2.infrastructure.primary;
+package tech.jhipster.lite.generator.server.springboot.mvc.security.oauth2.core.infrastructure.primary;
 
 import org.springframework.context.annotation.Bean;
 import org.springframework.context.annotation.Configuration;
-import tech.jhipster.lite.generator.server.springboot.mvc.security.oauth2.application.OAuth2SecurityApplicationService;
+import tech.jhipster.lite.generator.server.springboot.mvc.security.oauth2.core.application.OAuth2SecurityApplicationService;
 import tech.jhipster.lite.module.domain.resource.JHipsterModuleOrganization;
 import tech.jhipster.lite.module.domain.resource.JHipsterModulePropertiesDefinition;
 import tech.jhipster.lite.module.domain.resource.JHipsterModuleResource;
@@ -33,17 +33,5 @@ class OAuth2ModuleConfiguration {
       )
       .tags("server", "spring", "spring-boot", AUTHENTICATION)
       .factory(oAuth2::buildOAuth2Module);
-  }
-
-  @Bean
-  JHipsterModuleResource oAuth2AccountModule(OAuth2SecurityApplicationService oAuth2) {
-    return JHipsterModuleResource
-      .builder()
-      .slug("spring-boot-oauth2-account")
-      .propertiesDefinition(JHipsterModulePropertiesDefinition.builder().addBasePackage().build())
-      .apiDoc("Spring Boot - MVC - Security", "Add a account context for OAuth 2.0 / OIDC Authentication")
-      .organization(JHipsterModuleOrganization.builder().addModuleDependency("spring-boot-oauth2").build())
-      .tags("server", "spring", "spring-boot", AUTHENTICATION, "account", "user")
-      .factory(oAuth2::buildOAuth2AccountModule);
   }
 }
