+++ src/main/resources/generator/server/springboot/mvc/security/oauth2/core/docker/keycloak.yml.mustache
