+++ src/main/resources/generator/server/springboot/mvc/security/oauth2/core/main/infrastructure/primary/AuthenticatedUser.java.mustache
