+++ src/main/resources/generator/server/springboot/mvc/security/oauth2/core/test/infrastructure/primary/FakeRequestAttributes.java.mustache
